+++ src/Storage/SessionStore.php
@@ -11,4 +11,12 @@ interface SessionStore
      * @param $data
      */
     public function flash($name, $data);
+
+    /**
+     * Put data into the session.
+     *
+     * @param $name
+     * @param $data
+     */
+    public function put($name, $data);
 }