+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `sweet-alert` will be documented in this file.
 
+## v7.0.0 - 2023-04-20
+- Added confirmDelete function
+- Added themes feature
+- Upgraded SweetAlert2 to latest version
+- Bug Fixes
+- Docs Updated
+
 ## v6.0.0 - 2023-02-15
 - Added Laravel 10 Support
 - Bug Fixes